% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc_curve.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Calculate AUC}
\usage{
calculate_auc(ranked_scores, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{ranked_scores}{a data frame of scored sequences with FPR and TPR
columns}

\item{alternative}{\code{character(1)}. One of "two.sided, "less", or "greater".
The alternative hypothesis to test.}
}
\value{
a \code{data.frame} with columns \code{motif_id}, \code{auc}, \code{w_statistic},
\code{p_value}, and \code{alternative.}
}
\description{
Calculate the AUC of the ROC curve
}
\examples{
print("")
}
