% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sequences.R
\name{combine_scores}
\alias{combine_scores}
\title{Combine motif scores}
\usage{
combine_scores(
  scored_sequences,
  motif_groups,
  method = c("sum", "max", "mean")
)
}
\arguments{
\item{scored_sequences}{\code{data.frame}. The table of sequence scores to
combine.}

\item{motif_groups}{\code{list}. A named list of motif groupings to score.}

\item{method}{\code{character(1)}. One of "sum", "max", or "mean". The scoring
method to use. (Default: "sum")}
}
\value{
A \code{data.frame} with columns \code{sequence_name}, \code{motif_id}, and \code{score},
where the IDs in the  \code{motif_id} column are the names of the motif
groupings given in \code{motif_groups}.
}
\description{
Calculate a single score for each sequence based on multiple motifs.
}
\examples{
print("")
}
