% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_scored_sequences}
\alias{example_scored_sequences}
\title{Example table of sequence scores}
\format{
A data frame with 13,480 rows and 3 columns:
\describe{
\item{sequence_name}{The name of the promoter sequence.}
\item{motif_id}{The name of the motif.}
\item{score}{The score for this motif in this promoter sequence.}
}
}
\usage{
example_scored_sequences
}
\description{
A table containing 1,348 promoter sequences scored for 10 motifs.
}
\keyword{datasets}
