% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sequences.R
\name{get_sequences_with_motifs}
\alias{get_sequences_with_motifs}
\title{Get genes with motifs}
\usage{
get_sequences_with_motifs(scored_sequences, motif_ids, min_score = 1)
}
\arguments{
\item{scored_sequences}{a data frame of FIMO results}

\item{motif_ids}{the IDs of the motifs to look for}

\item{min_score}{the minimum score to count a motif occurrence}
}
\value{
a data frame of sequences that have the given motifs
}
\description{
Get a list of genes that contain one or more given motifs
}
\examples{
print("")
}
