% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sequences.R
\name{normalize_scores}
\alias{normalize_scores}
\title{Normalize sequence scores}
\usage{
normalize_scores(scored_sequences, bed_file)
}
\arguments{
\item{scored_sequences}{\code{data.frame}. The table of sequence scores to
normalize.}

\item{bed_file}{\code{character}. The path to the BED format file of the
sequences.}
}
\value{
A \code{data.frame} with columns \code{sequence_name}, \code{motif_id}, and \code{score}.
}
\description{
Normalize sequence scores based on the lengths of the sequences by performing
a linear regression.
}
\examples{
print("")
}
