% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc_curve.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot ROC curves}
\usage{
plot_roc_curve(ranked_scores)
}
\arguments{
\item{ranked_scores}{a data frame of scored sequences with FPR and TPR
columns}
}
\value{
a ggplot object with the ROC curve
}
\description{
Plot a ROC curve
}
\examples{
print("")
}
