% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo.R
\name{preprocess_fimo_results}
\alias{preprocess_fimo_results}
\title{Preprocess FIMO results}
\usage{
preprocess_fimo_results(
  fimo_results,
  sequences = NA,
  motifs = NA,
  max_pvalue = 1e-04
)
}
\arguments{
\item{fimo_results}{\code{data.frame}. The table of FIMO results to process.}

\item{sequences}{\code{character} or \code{NA}. The list of all sequences that should
be in the final table. If \code{NA}, all of the sequences in the input
\code{fimo_results} table that have at least one motif that passes the p-value
threshold are used. (Default: NA)}

\item{motifs}{\code{character} or \code{NA}. The list of all motifs that should be in
the final table. If \code{NA}, all of the motifs in the input \code{fimo_results}
table that have at least one occurrence that passes the p-value threshold
are used. (Default: NA)}

\item{max_pvalue}{\code{numeric(1)}. The maximum p-value to consider. Set to \code{1}
to skip filtering by p-value. (Default: 1e-4)}
}
\value{
A \code{data.frame} with columns \code{motif_id}, \code{motif_alt_id},
\code{sequence_name}, \code{start}, \code{stop}, \code{strand}, \code{score}, \code{p_value}, \code{q_value},
and \code{matched_sequence}.
}
\description{
Prepare FIMO results for downstream analyses by filling in any missing
combinations of sequences and motifs and reformatting the table if necessary.
}
\examples{
print("")
}
