% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sequences.R
\name{rank_scores}
\alias{rank_scores}
\title{Calculate TPR and FPR}
\usage{
rank_scores(scored_sequences_fg, scored_sequences_bg)
}
\arguments{
\item{scored_sequences_fg}{\code{data.frame}. Sequences scored by motif
occurrences.}

\item{scored_sequences_bg}{\code{data.frame}. Sequences scored by motif
occurrences.}
}
\value{
A \code{data.frame} of scored sequences with FPR and TPR columns added
}
\description{
The title is good enough for now
}
\examples{
print("")
}
